feat: ✨ use ts local action
