+++ .github/actions/app-token/package.json
@@ -2,9 +2,10 @@
   "name": "app-token",
   "version": "1.0.0",
   "private": true,
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "start": "node index.js"
+    "clean": "rimraf lib",
+    "build": "yarn clean && tsc"
   },
   "dependencies": {
     "@actions/core": "^1.2.5",
@@ -12,8 +13,10 @@
     "@octokit/app": "^4.2.1",
     "is-base64": "^1.1.0"
   },
-  "devDependencies":{
+  "devDependencies": {
     "@types/is-base64": "^1.1.0",
-    "@types/node": "^14.6.4"
+    "@types/node": "^14.6.4",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.0.2"
   }
 }
